contracts: move ContractPermissionDescriptor into manifest
